+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Added Url support to event tooltips (can be clicked, which opens the Url in a new window/tab).
 - Only valid Urls can now be entered in the Event Editor dialog (an error message is shown for invalid entries).
 - Events that contain an invalid Url when added/updated via public functions are now cleared.
+- Added new setting "allowHtmlInDisplay", which allows HTML to be used in the display (defaults to false).
 - 
 
 ### Version 1.4.5: